Code 3
In this code, the while loop continues as long as the number(x) is less than or equal to 100. Inside the loop, it checks if the current number is not divisible by 3 and not divisible by 5 using the modulo operator (%). If the condition is True, it prints the number. Finally, the number is incremented, and the loop continues until number exceeds 100